+++ build.gradle
@@ -35,6 +35,7 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     implementation 'org.springframework.boot:spring-boot-starter-webflux'
     implementation 'org.springframework.boot:spring-boot-starter-websocket'
+    implementation 'org.springframework.boot:spring-boot-starter-mail'
 
     // Lombok
     annotationProcessor 'org.projectlombok:lombok'
@@ -61,6 +62,9 @@ dependencies {
 
     // Hibernate-spatial
     implementation 'org.hibernate:hibernate-spatial:6.6.3.Final'
+
+    // Hibernate-util
+    implementation 'io.hypersistence:hypersistence-utils-hibernate-63:3.9.2'
 
     // Mapstuct
     implementation 'org.mapstruct:mapstruct:1.6.3'
@@ -95,9 +99,6 @@ dependencies {
     // Amazon S3
     implementation 'software.amazon.awssdk:s3:2.30.3'
 
-    // Monitoring
-    implementation 'org.springframework.boot:spring-boot-starter-actuator'
-
     // Swagger + REST Docs
     testImplementation 'org.springframework.restdocs:spring-restdocs-mockmvc'
     testImplementation 'com.epages:restdocs-api-spec-mockmvc:0.19.4'
@@ -108,8 +109,8 @@ dependencies {
     runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.12.3'
     runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.12.3'
 
-    // groovy
-    implementation 'org.codehaus.groovy:groovy-all:3.0.9' // Groovy 의존성 추가
+    //groovy
+    implementation 'org.codehaus.groovy:groovy-all:3.0.9'
 }
 
 // JaCoCo 버전 설정
